uint & size__t -> unsigned for Vector.2

+++ src/amun/common/histories.cpp
@@ -207,9 +207,9 @@ Hypotheses Histories::GetSurvivors() const
   return ret;
 }
 
-std::vector<size_t> Histories::Hypo2Batch() const
+std::vector<unsigned> Histories::Hypo2Batch() const
 {
-  std::vector<size_t> ret;
+  std::vector<unsigned> ret;
   for (size_t i = 0; i < coll_.size(); ++i) {
     const HistoriesElementPtr &ele = coll_[i];
     if (ele) {
